Fixed incorrect top-of-VM-stack detection in XRETURN
XRETURN would check the stack frame for a NULL function pointer,
but since some recent VM changes, what actually happens is you
get a 0 base index - and trying to find a stack frame there will
result in a segfault. Now it checks for that instead, just like
the RETURN instruction.