Moving to CMake, and merging eelbox into the main project
The EEL core and the "core" modules are now build as one
library, and eelbox as another.

Only one executive, 'eel' is built, and it's actually the
old 'eelbox'. We no longer build a non-eelbox executive,
except for 'eeltest' (which is updated and working again),
but that's never installed.

Static libraries only, for now! Need to get Find*.cmake
and/or pkg-config files in order again, and it's probably
not a great idea to build shared libraries anyway, until
the API and ABI has stabilized.

Closes #2.