Fixed colorkey transparency in OpenGL UploadTexture()
When deciding whether or not to use an alpha channel in the GL
texture, UploadTexture() only checked for an actual alpha channel
in the source surface. Thus, anything that loaded as a colorkeyed
SDL surface would have the transparency information dropped at
this point! Now, we also check the flags for SDL_SRCCOLORKEY.

Added some read-only fields to the SDL_Surface wrapper while
debugging this.

Closes #10.