Temporary fix for the load() segfault issue
I think I've figured out what's causing it: Some problem with RETX*
into native functions.

EEL implements exception handling by compiling 'try' and 'except'
blocks as functions, which makes it easy to cleanly exit them at
any point, and that's also how the 'retry' feature could be
implemented so trivially.

However, when executing a 'return' statement inside an exception
block, we can't use the RETURN instruction, as that would just leave
the exception block. So, we throw a special exception using one of
the RETX* instructions.

Now, somewhere around here is where the problem is; that mechanism
is somehow failing to handle the differences between native functions
and EEL functions. If we return directly to a C function (which is
the case when builtin.load() returns to eel_load() in e_module.c),
something goes wrong and we get corrupt return values.

This "fix" simply wraps load() with a trivial EEL function, so that
we avoid the RETX* into native code for now.