Started using EEL_MODULE_DIR
EEL_MODULE_DIR is now based on CMAKE_INSTALL_PREFIX, so on Un*x,
the "official" modules are typically installed in /usr/lib/eel/.

Added the serialize.eel EEL (JSON-like) and binary data
(de)serialization module from Kobo II and Infilterion.

serialize() now has an optional flag SERIALIZE_DROP_UNKNOWN that
makes it silently drop data of unsupported types. The normal
behavior is now to throw an exception. (It used to print or log
a warning message in that case.)

Added "./modules" to $.path_modules, so EEL modules can be placed
in a local directory, so Windows "bundle style" distributions and
the like can run without the modules being installed in the system.

Moved EBGUI (the old SDL 2D GUI toolkit) into a subdirectory eelium
in the modules directory, and updated the tests/demos accordingly.