'import' can now find local modules regardless of current directory
Added $.current_module_path, which holds the path to the module currently
being compiled, whenever available.

__load_via_path_modules() now tries $.current_module_path before going
through $.path_modules.

When no module is being compiled, $.current_module_path holds the path to
the first module compiled in the state; typically the main application
module. This allows runtime load() calls (which also serves as the backend
for 'import') to behave more reliably - though it should be noted that a
module imported from a subdirectory using load() will still get the path
of the main module; NOT the path of the module itself.

Closes #40.