Added __modpath export to compiled modules
When a module is compiled, the formerly temporary path that calculated
for $.current_module_path is now exported along with __modname etc, so
code in the module can retrieve it with __exports().__modpath.

This is useful in modules that want to load files (images, sounds etc)
using paths related to the location of the module itself.