Added Lua/JSON style table constructor shorthand syntax
EEL now allows "{ ... }" as shorthand for "table [ ... ]", similar to the
syntax of Lua tables and JSON objects.

Added the new syntax to the "eel" format of serialize(), along with a new
flag SERIALIZE_NO_SHORTHAND to disable shorthand syntax for both tables
and arrays.

Updated tests scripts and modules to the new shorthand table syntax.

Closes #61.