Fixed Eelium module initialization error handling
There isn't going to be anything of interest from the compiler error log
for eel_perror() to print if one of the Eelium modules fail to initialize,
so the 'verbose' flag has been removed.

Instead, eelium_init_bindings() will print the name and description of the
exception code returned, of one of the modules fail to initialize.

Closes #73.