Implemented proper 'in' operator for table, string and dstring
The 'in' operator is now backed by a new metamethod EEL_MM_IN, which is
currently implemented by the table, string and dstring types.

Tables try normal key look-up as before.

String and dstring search their buffers, and they support both single
characters (any value that can be cast to an integer) and substrings in
the form of other strings or dstrings.

Closes #82.