Added cmdline.eel module
This handy utility module deals with command line arguments and similar.
At this point, it exports a procedure usage() that prints documentation,
and a function parse() that parses and checks arguments. Both take
argument specifications in the same format, which allows arguments to be
named, typed and provided with default values and descriptions.