Some MXE and MingW64 build fixes
The master branch of MXE has moved on to MingW64 for both 32 and 64 bit
Windows builds, so some updates were needed;

Target changed from i686-pc-mingw32 to i686-w64-mingw32.shared. That is,
we build DLLs now!

Linking with winmm, for the timer calls. (This probably leaked in from
SDL when linking statically.)

When building for MingW64, staying away from _setjmp(), which has been
replaced with some weird version with two arguments, and _longjump(),
which is apparently missing in msvcrt, despite being declared.

And, a workaround for this mysterious CMake problem, where find_package()
for a certain library only works once in a project... In this case,
PkgConfig.