Fixed some argument checking bugs in the Eelium SDL module
SetAlpha() and SetColorKey() no longer ignore the flags argument when the
alpha/colorkey argument is specified.

MapColor() now throws an error when given an unsupported argument count.

EventState() uses the correct argument value for the 'state' argument.

Refactor/fix in esdl_PlayAudio(); ALWAYS use '>=' when simply checking
whether or not an argument was specified!