Added Eelium SDL.FilterPixel() with utilities
Added SDL.FilterPixel(), which takes a fixed point filter core in the form
of a vector_s32, with width and height as the first two items. This is not
terribly optimized (brute force per-pixel clipping!), but still "slightly"
faster than the all EEL "average" filter in TIPE...

Added gfxutils.NormalizeFilterCore() and CircularFilterCore() utility
functions, that normalize/scale, and generate filter cores compatible with
SDL.FilterPixel(). CircularFilterCore() takes two extra arguments that
allow the weights to be tapered as desired from the edge to the center of
the circular body of the filter core.

Closes #100.