Hide classes exported from modules from the root namespace
Classes exported from modules via eel_export_class() are now given "invalid"
symbol names in the VM class registry, so that the compiler can only find them
in the correct place: In the 'export' table of the module.

This should also allow multiple modules to export unique classes with the same
name, as long as the modules have different names.

The reason the symbols still have names at all in the registry is that a fair
bit of refactoring would be necessary to get dumps, warning/error messages etc
to work properly without them.

Closes #120.