Fixed indexed color case in SDL.Surface constructor
While the SDL_CreateRGBSurface() states that "If depth is 4 or 8 bits, an empty palette
is allocated for the surface," it seems that the mask need to be zero as well. Have not
checked whether or not SDL2 actually creates an RGB232 surface, but it does NOT create
an empty palette unless the masks are also zero.

Closes #136.